+++ ejo-core.php
@@ -3,7 +3,7 @@
  * Plugin Name:         EJO Core
  * Plugin URI:          http://github.com/ejoweb/ejo-core
  * Description:         EJOweb core functionalities for theme development. Including some nifty debug tools.
- * Version:             0.9.1
+ * Version:             0.9.2
  * Author:              Erik Joling
  * Author URI:          http://www.ejoweb.nl/
  * Text Domain:         ejo-core
@@ -28,7 +28,7 @@
 final class EJO_Core 
 {
     /* Version number of this plugin */
-    public static $version = '0.9.1';
+    public static $version = '0.9.2';
 
     /* Holds the instance of this class. */
     protected static $_instance = null;
@@ -158,6 +158,9 @@ final class EJO_Core
 
         /* Cleanup Backend */
         require_if_theme_supports( 'ejo-cleanup-backend', EJO_DIR . 'includes/unregister-widgets.php' ); // Widget Unregistering
+
+        /* Admin Image Select Script */
+        require_if_theme_supports( 'ejo-admin-image-select', EJO_DIR . 'includes/admin-image-select.php' );
      
     }